Database Cache Housekeeping
Introduce a Housekeeper to the database cache. It performs its tasks
asynchronously in its own thread to avoid blocking the main one.

**Aim**
The hope is that this will create a self managing LRU cache, with the
cache size managed by setting a minimum record age.

It's activity should scale with cache usage so we avoid eating resources
on inactive processes.

Communication is done via a SizedQueue, which will prevent memory bloat
if the housekeeping thread is starved.

Cache management is best effort or losing messages or shutting down
before all tasks are completed is ok.

It has two tasks:
1. Touching read records
2. Deleting old records

**Touching read records**

When the cache does a read a task is queued to touch the records that
were read. We want this so we can implement a LRU algorithm for
deletions.

The Housekeeper batches up those deletions.

**Deleting old records**

When records are written to the cache we queue a task for the
Housekeeper to delete some records. Currently this is set to delete one
more record than was written to allow the cache size to shrink if
necessary.

The cache is configured with a field to use to find old records. It
could be created_at for FIFO, updated_at for LRU or expires_at for fine
grained control, but the field should be indexed. It also has a
minimum age for record deletion.

Things that still need to be worked on:
- Thread management - priority, stopping, working across forks
- Thread error handling/reporting
- Prevent deletion thrashing - the deletion tries to delete the N oldest
records so if we had lots of processes doing this at the same time it
might cause issues. Also need to see what happens when there are almost
no records to delete.
- Configurablity/defaults