Touch records asyncronously (#17)
We need to touch read records for an LRU cache. However we also want to
avoid writing to the records every time we to a cache read.

To handle this we add a `AsyncronousExecutor` that can deal with this
and future async tasks (such as deleted old/expired records).

Internally it uses a `Concurrent::SingleThreadExecutor`. It batches up
the ids and touches them every `touch_batch_size` records.

This is best effort - we don't attempt to flush the records on shutdown
so not everything is guaranteed to by marked.