Trim the cache (#18)
* Trim the cache

Trim the cache to prevent uncontrolled growth. The deletions happen on
the background thread.

The configuration for this is:
- trim_batch_size - how many records to trim at once
- trim_min_age    - the minimum age of records before they can be trimmed
- cache_full      - a boolean or callable that indicates that the cache is full

Every time we write to the cache we will report that to the record
trimmer. The trimmer will apply a multiplier to that number to ensure we
are attempting to delete slightly more data than we are adding.

When the counter exceeds `trim_batch_size` we'll attempt to delete a
batch of that size filtering out records that are newer than the
`trim_min_age`.

`cache_full` is a boolean/Proc that defaults to false. If it is a Proc,
we check it before deleting records. If it reports that the cache is
full we'll ignore the min age when deleting to ensure we are actively
reducing the cache's size.

`cache_full` could be implemented as a check on the table statistics if
required.

Deleting records could be `DELETE from cache order by updated_at limit 100`
but that's likely to cause issues if we have a lot of concurrency.

Instead we do:
```
SELECT id from cache order by updated_at limit 500;
DELETE from cache where id = <random selection 100 of those 500 ids>;
```

This does mean we are not guaranteed to delete 100 records if a
concurrent delete has deleted some of those.

We'll need to see how it operates in practice and tune the multiplier
parameters to ensure that there are sufficient deletions without too
much contention.

* Use port 33060