Update counters synchronously (#44)
Move updating the trim counters to the same calling thread. Only the
actual trimming will be done asynchronously which should reduce the
amount of work done on the async thread significantly.

Now uses an atomic fixnum as we no longer do all the modifications in
a single thread.

Uses compare and set to reduce the counter and trigger trimming. If the
atomic fixnum is updated and compare and set returns false, it doesn't
retry, assuming that the concurrent thread that updated the counter will
trigger trimming instead.