Use maglev hashing
Compared to a hash ring [Maglev hashing](https://static.googleusercontent.com/media/research.google.com/en//pubs/archive/44824.pdf)
is faster, fairer and uses less memory with a table size of 2053.

The table size must be a prime and is an upper limit on the number of
nodes begine hashed.

Larger table sizes are fairer but require more memory.