Use a dedicated thread for each connection
Switching the connection out when we need to change shard is expensive.
It requires checking a connection out of the pool. That will verify the
connection which requires a round trip to the database.

Instead we'll create a dedicated thread for each connection and run
the query on it via a `Concurrent::Promise`.

This has two benefits:
1. There's no connection switching required as each of the threads only
talks to one shard.
2. Multi reads and writes can now happen in parallel as we trigger them
as before getting the results from the promises.

The existing async operations (trimming and writing to non primary
clusters) don't use these executors, they have their own one instead
that does the connection switching. We don't need to worry so much about
that as they are not time critical operations, but there's some scope
here to tidy things up as they have a lot in common with the new
threaded operations.