Expire more for larger writes
Expiries should be proportional to the number of items written, not the
number of write operations.

We multiply the number of writes by the expiry chance to determine how
many expiries to trigger. We trigger an expiry for every whole number in
that muliplied value and determine whether to trigger an extra expiry
based on the remainder.