Support installing with SQL schema format
As with the schema format, we'll just add a structure file containing
the SQL for creating the cache tables as this will work with
`bin/rails db:prepare` even if you only run the cache database in
production.

We bundle the SQL files for MySQL, PostgreSQL, and SQLite - they are
generated with `bin/generate_sql_schemas`.